refactor: rearrange type imports + eslint rule
