+++ src/components/icons/icon-bits.tsx
@@ -1,4 +1,4 @@
-import { type SVGAttributes } from 'react';
+import type { SVGAttributes } from 'react';
 
 export function IconBits(props: SVGAttributes<SVGSVGElement>) {
   return (
