+++ src/components/icons/icon-chart.tsx
@@ -1,4 +1,4 @@
-import { type SVGAttributes } from 'react';
+import type { SVGAttributes } from 'react';
 
 export function IconChart(props: SVGAttributes<SVGSVGElement>) {
   return (
