+++ src/components/icons/icon-follow.tsx
@@ -1,4 +1,4 @@
-import { type SVGAttributes } from 'react';
+import type { SVGAttributes } from 'react';
 
 export function IconFollow(props: SVGAttributes<SVGSVGElement>) {
   return (
