+++ src/modules/pre-stream-scene/index.astro
@@ -1,7 +1,7 @@
 ---
 import Document from '../_document.astro';
 import { StreamTitle } from './components/stream-title';
-import { type PreStreamVariants } from './types';
+import type { PreStreamVariants } from './types';
 
 type Props = {
   headerText: string;
