+++ src/server/modules/deepdip/get-global-leaderboard/types.ts
@@ -14,19 +14,3 @@ export interface LeaderboardResult {
   /** Player's selected colour in-game (in RGB percentage) */
   color: number[];
 }
-
-const DD2_API_ENDPOINT = 'https://dips-plus-plus.xk.io';
-
-export async function getGlobalLeaderboard() {
-  const response = await fetch(`${DD2_API_ENDPOINT}/leaderboard/global`);
-  const data: LeaderboardResult[] = await response.json();
-
-  return data;
-}
-
-export async function getCurrentProgress(accountId = '15d23e07-07ac-4093-bbfd-28d393daf0c0') {
-  const response = await fetch(`${DD2_API_ENDPOINT}/leaderboard/${accountId}`);
-  const data: LeaderboardResult = await response.json();
-
-  return data;
-}
