+++ src/server/modules/notion/get-current-stream/types.ts
@@ -1,0 +1,8 @@
+export type CurrentStreamInformation = {
+  id: string;
+  date?: string;
+  series?: string;
+  categories?: string[];
+  stream_name?: string;
+  description?: string;
+};
