+++ src/server/modules/notion/get-overlay-data/types.ts
@@ -1,0 +1,5 @@
+export interface OverlayDataInformation {
+  breakReturnTime?: string;
+  timeSignal?: string;
+  streamStart?: string;
+}
